requirements.txt
# requirements.txt

apache-airflow==2.8.2
apache-airflow-providers-standard
structlog==24.1.0
SQLAlchemy==1.4.49
requests>=2.23.0,<2.27.2
PyYAML==6.0
psycopg2-binary==2.9.6
python-json-logger==2.0.7
redshift-connector==2.0.907
